Emit an error when proof_for_contract function is not found (#3609)
Currently, Kani panics if the function specified in the
`proof_for_contract` attribute is not found (e.g. because the function
is not reachable) (see #3467). This PR adds an error message pointing
out the issue.

Towards #3467 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.