[Lean] Rename user-facing options from Aeneas to Lean (#3630)
This is a follow-up on #3514. As @celinval suggested
([here](https://github.com/model-checking/kani/pull/3514#discussion_r1790949304)),
renaming all user-facing options from "Aeneas" to Lean.

Inside the Kani compiler which is only concerned with producing LLBC
(and doesn't know about Lean/Aeneas), I'm renaming the relevant feature
from `aeneas` to `llbc`, and the backend is now called the LLBC backend
as opposed to the Aeneas backend.

Towards #3585 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.