Add regular & fixme tests for function contracts (#3371)
I'm adding a few fix-me tests that I bumped into while working on #3363.
Most of them will be fixed by #3363, except the one related to #3370.

The original PR is already quite large, so I decided to just push all of
these as fixme tests for now. This is now ready for review!

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Felipe R. Monteiro <felisous@amazon.com>
Co-authored-by: Zyad Hassan <88045115+zhassan-aws@users.noreply.github.com>
Co-authored-by: Carolyn Zech <cmzech@amazon.com>