Fix loop contracts transformation when loops in branching (#3640)
In loop contracts transformation, we want to visit non-loop blocks
before visiting loop blocks. So we pop `loop_queue` only when `queue` is
empty. We should use `or_else` instead of `or` which always pop the
front of `loop_queue`.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.