Add a timeout option (#3649)
This PR adds a new option, `--cbmc-timeout` which can be used for
setting a timeout for the CBMC run.

The timeout applies on a per-harness basis in the Kani run.

The implementation uses [tokio ](https://tokio.rs/) to perform
non-blocking reads from the CBMC process's `stdout` so that the timeout
can be checked asynchronously.

**Call-outs:**

There are other names I considered for the option:
- `--harness-timeout`: This makes it clearer that the timeout applies to
each individual harness (and not to the overall Kani run). It doesn't
indicate that the timeout only applies to the CBMC portion of the flow
though.
- `--engine-timeout`: This makes it a bit more general across multiple
engine/backends. The notion of an engine is not really part of the Kani
jargon currently though, so might not be clear to users.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.