codegen: Ask the layout if it is uninhabited, not its impl detail (#3675)
Directly question the layout if it is uninhabited. This makes Kani more
resilient to changes in the upstream rustc toolchain, as the convenience
function which exposes this is precisely to allow compilers to not
fixate as much on the innermost details of rustc's codegen process.

Found while reviewing https://github.com/model-checking/kani/pull/3671