Derive `Arbitrary` for enums with a single variant (#3692)
Current implementation of deriving macro for arbitrary trait produces
code that fails with a compilation error when invoked on a enum with a
single variant due to type ambiguity in `match` expression scrutinee.
This patch adds special handling of such enums, eliminating the
redundant match and fixing the error.

Resolves #3691

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.