Remove symtab json support (#3695)
I believe the plan was to keep this in case of a regression for about a
year. The time has long gone, so I think it's time to get rid of it.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.