Apply loop contracts only if there exists some usage (#3694)
In this PR, we detect the usage of loop contracts, and apply loop
contracts with goto-instrument only if there exists some usage of loop
contracts in the codegen unit.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.