Remove CBMC viewer and visualize option (#3699)
This builds upon @jaisnan's changes in
https://github.com/model-checking/kani/pull/3302.

The PR removes the `--visualize` option and the CBMC viewer from the
Kani installation and CI.

There is still some room for clean-up, specifically with the python
dependency. This can be handled in a separate PR.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Jaisurya Nanduri <jaisnan@amazon.com>