Ignore derivative in Cargo deny (#3708)
`derivative` is marked as unmaintained by RUSTSEC but still used in
Charon. We exclude it from the deny check until `derivative` is replaced
in Charon (https://github.com/AeneasVerif/charon/pull/459).

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.