Fix codegen for rvalue aggregate raw pointer to an adt with slice tail (#3644)
Fix the codegen for a raw pointer to a dynamically-sized ADT. Traverse
the type until we find the tail, which will either be a trait object or
a slice, and generate the pointer accordingly.

Resolves #3638, #3615

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Celina G. Val <celinval@amazon.com>
Co-authored-by: Qinheping Hu <qinhh@amazon.com>