List Subcommand Improvements (#3729)
Improvements to the list subcommand.

- Print the location of the JSON file for `--format=json` (#3633)
- Use `CommonArgs` instead of `VerificationArgs` (#3632)
- Add a Markdown option. The pretty table is nice for smaller packages,
but as we progress with the standard library verification effort, it's
too massive to be easily readable. In this case, it's better to print a
markdown table and then just read it in a markdown editor. (Once this PR
is merged, the [kani list workflow
step](https://github.com/model-checking/verify-rust-std/pull/146) will
use the markdown option).

Resolves #3632, #3633

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.