Upgrade toolchain to 2024-11-28 (#3753)
Update parameter and function indices in LLBC tests (again) that appear
to have changed without any of the upstream changes obviously causing
this. Those tests will eventually be removed in upcoming changes.

Resolves: #3752

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.