Add out of bounds check for `offset` intrinsics (#3755)
Note that the `offset` intrinsics is lowered to `Rvalue::Offset`. Thus,
we have to replace the entire statement.

This PR does not address `ptr_offset_from` and
`ptr_offset_from_unsigned`. I'll push those changes to a separate PR.

Resolves #1233 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Carolyn Zech <cmzech@amazon.com>