Update charon submodule (#3801)
Update Charon submodule to
https://github.com/AeneasVerif/charon/commit/adc0a855aa1428d68e4270edef0d52131cef06ab

Relevant Charon PRs:
https://github.com/AeneasVerif/charon/pull/457: This required updating
the code that creates places to also pass in the type.
https://github.com/AeneasVerif/charon/pull/464: This replaced the
file-to-id hash table by a vector, thus requiring that we add one in
Kani's translation context.
https://github.com/AeneasVerif/charon/pull/474: The translation context
is no longer parameterized by a lifetime
https://github.com/AeneasVerif/charon/pull/491: This required changing
the type of indices

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: thanhnguyen-aws <ntson@amazon.com>