Generate contracts of dependencies as assertions (#3802)
Add an unstable option `-Zcontracts-as-assertions` that generates
contract clauses as assertions. This works by adding a new contract mode
that outputs the original body of the function with its contract clauses
asserted. During the `FunctionWithContractPass`, if this option was
passed, we use this closure instead of the original body of the
function.

I'm omitting a description of why this is useful from the PR
description--there's one in the module documentation, and I want that to
be a self-contained explanation (i.e., if something is unclear, I want
to fix it there).

(Also did a minor bit of cleanup, renaming `ClosureType` to
`ContractMode` and moving `split_for_remembers` to the shared module
since it has contract-specific logic.)

Resolves #3326

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Zyad Hassan <88045115+zhassan-aws@users.noreply.github.com>