Fix validity checks for `char` (#3853)
Ensure value validity checks for `char` in Kani matches the Rust
documentation:

> A char value must not be a surrogate (i.e., must not be in the range
  0xD800..=0xDFFF) and must be equal to or less than char::MAX.

The existing code was relying on the compiler ABI information, which can
only express one continuous value range for value validity. But `char`
is a special case, and the Rust compiler understands the gap in valid
`char`. This change makes Kani compiler aware of this gap too.

Resolves #3241

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.