Stub linker to avoid missing symbols errors
Kani compiler does not generate real object files when compiling a crate
and its dependencies for verification. Instead, the compiler generates
goto binaries. This can cause linking errors. Instead, stub it with
echo.

We should consider creating a stub linker instead, but echo will do it
for now.

Note that this change only applies to target artifacts, not those
compiled for the host, such as build dependencies and build scripts.