Stub linker to avoid missing symbols errors (#3858)
Kani compiler does not generate real object files when compiling a crate
and its dependencies for verification. Instead, the compiler generates
goto binaries. This can cause linking errors. Instead, stub it with
echo.

We should consider creating a stub linker instead, but echo will do it
for now.

Note that this change only applies to target artifacts, not those
compiled for the host, such as build dependencies and build scripts.

Resolves #3817 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.