remove flag float-overflow-check (#3873)
Resolves #3620

Removing default flag float overflow check, which reports overflow for
arithmetic operations over floating-point numbers that result in +/-Inf.
This doesn't panic in Rust and it shouldn't be consider a verification
failure by default.

Users can enable them using --cbmc-args as and when required.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.