Support concrete playback for arrays of length 65 or greater (#3888)
### Problem
When CBMC generates a JSON trace for a nondeterministic array, it will
output a key-value pair `elements: [array]`, where `array` contains
concrete values for each element in the array. If the array is length 64
or shorter, it will _also_ output the elements of the array as separate
values in the trace (so each element is in the trace twice). Prior to
this PR, concrete playback relied on the latter output format to find
elements of an array. However, when the array was length 65 or greater,
those elements wouldn't be values in their own right, so we wouldn't
find any values for the array and just output an empty array.

### Solution
This PR changes our representation of concrete values to handle arrays
explicitly; i.e., to look for the `elements` array and populate the
concrete values of the array from that instead.

### Callouts
For those wondering why the `playback_already_existed` test changed,
it's because we're hashing `ConcreteItem` instead of `ConcreteValue`, so
the hash changes.

Resolves #3787

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.