Autoharness: Don't panic on `_` argument and add `_autoharness` suffix to GOTO files (#3942)
Turns out that if an argument is just `_`, it doesn't have
`var_debug_info` in StableMIR. So this PR changes autoharness behavior
to provide `_` for the argument name when `var_debug_info` is None
(instead of panicking).

I don't currently know of any other cases where there wouldn't be
`var_debug_info`, but we should pay extra attention to the cases where
autoharness reports `_` as an argument name and update this logic if
need be.

I also added an `_autoharness` suffix to the generated GOTO files so
that it's easier to tell which files are from autoharness and which
aren't.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.