Autoharness: metadata improvements and enable standard library application (#3948)
Update the autoharness subcommand to:

- Print the types of the arguments that don't implement Arbitrary
- Print a table of the functions we chose to verify after codegen
finishes, and move the printing of the functions we skipped to be after
codegen as well, instead of after verification. This lets us iterate
faster to debug because we can pass `--only-codegen` and see
pretty-printed data instead of waiting for verification to finish.
- Add a `--std` flag which, when provided, runs autoharness against the
provided path to the standard library.

Towards #3832

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.