Upgrade toolchain to 2025-03-18 (#3959)
The main change to get the toolchain upgrade to 2025-03-18 to work is to
make this change in `kani-regression.sh`:
```diff
-RUSTFLAGS="-D warnings" cargo build --target-dir /tmp/kani_build_warnings
+RUSTFLAGS="-D warnings" cargo build --target-dir /tmp/kani_build_warnings --no-default-features --features cprover
```
Explanation:
1. `cargo build` suppresses warnings from dependencies unless they're
local dependencies (e.g. included with `path = ...`) (see
https://github.com/rust-lang/cargo/issues/8546)
2. Since we integrate `charon` as a git submodule and add it as a
dependency using `path = ...`, its warnings are not suppressed
3. The toolchain upgrade includes adding a `#[must_use]` attribute on an
enum used in charon:
https://github.com/rust-lang/rust/commit/2439623278. This results in a
warning that causes the `RUSTFLAGS="-D warnings" cargo build` to fail:
https://github.com/model-checking/kani/actions/runs/13914318184/job/38934470822#step:4:2983

The short-term solution in this PR is to exclude the Charon feature when
building with `-D warnings`.

The long-term solution is to fix the warning in upstream Charon (if it's
not already fixed), and update our Charon pin to point to a commit that
includes the fix. Updating the Charon pin requires us to upgrade our MIR
to ULLBC module though, which will require a non-trivial amount of work
since it's 2.5 months old.

Resolves #3944

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.