Remove unstable-features from code formatting script (#3962)
The code formatting script
[scripts/kani-fmt.sh](https://github.com/model-checking/kani/blob/main/scripts/kani-fmt.sh)
currently uses `--unstable-features` to enable the usage of the
`--ignore` pattern in
[rustfmt.toml](https://github.com/model-checking/kani/blob/main/rustfmt.toml).
Since we're already using a find command to find all the files to
format, it is easy to instead add the patterns to exclude to the find
command rather than relying on an unstable feature.

Resolves #ISSUE-NUMBER

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.