Fix `autoharness` termination test & print metadata in alphabetical order (#3971)
https://github.com/model-checking/kani/pull/3968 [failed in the merge
queue](https://github.com/model-checking/kani/actions/runs/14180398658/job/39724995535)
because the `cargo_autoharness_termination` test assumed that the
recursion test wouldn't time out, but it did. This doesn't happen often,
since this test has worked up until now, but in the interest of having
robust tests, this PR splits the termination test into two:

- `autoharness_timeout`, which tests that the default timeout of 60s
kicks in
- `autoharness_unwind`, which tests that the default unwind bound of 20
kicks in

The unwind test specifies a timeout of 5m to ensure that the unwind
bound takes effect before the timeout.

While I was fixing this, I also fixed some inconsistencies with the
printed metadata:
- We use "Chosen Function" in one table and "Selected Function" in
another--changed it to "Selected Function" in both to be consistent
- The functions aren't printed alphabetically in 2/3 tables--fixed so
that they're always sorted

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.