Add an option to skip codegen (#4002)
This PR adds a new unstable option, `--no-codegen`, that can be used for
running Kani, just to check if the code compiles, i.e. similar to
running `cargo check`. This is useful as a lightweight check for
compilation errors including code under `#[cfg(kani)]`.

This is implemented through passing rustc's `-Zno-codegen` option, which
is what `cargo check` passes to `rustc`.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Carolyn Zech <cmzech@amazon.com>