Autoharness: Harness Generation Improvements (#4017)
The important changes are:

- 37db951c921686a7a5f92833c65e74f745c68eb0, which will allow autoharness
to eventually work with modifies clauses. Since we don't generate
autoharnesses for functions that take pointers as arguments anyway, this
doesn't affect current functionality.
- a325fd8270d97adb2e15ffec1cc365a34c597540, which ensures that the
`--function` passed to CBMC is the mangled name of the automatic harness
intrinsic, not the target function

Towards https://github.com/model-checking/kani/issues/3832

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.