Autoharness: Update Filtering Options (#4025)
## Summary
- fda814bcdfadec64b570c9362693f6ea92bd009e: Make the autoharness filter
flags work on the standard library by moving them to
`kani_compiler_flags`, which ensures they're passed to all Kani compiler
invocations.
- c0430ec8aecb827ac29d81916081c91dc95665e6: Print the crate name in our
output tables.
- 093fc6b57b56e8bac9f5fb3baa8a49b8c586f7b2: **Breaking Change** to
rename `--include-function` and `--exclude-function` to mention
`pattern`s instead, which makes it clearer that they talk about
substrings of the total paths (e.g., modules). Also implement the
suggestion from
https://github.com/model-checking/kani/pull/3922#discussion_r1982170163
so that the flags are no longer mutually exclusive.
- 9e35fcab81ef0dcf8c0f01441b8b1a753948bdd8: Let the above flags filter
on crate names as well.
- 66444e40d7d12ab7ad1b118cf814f17206be9ee4: Warn if an exclude flag
makes an include flag moot.

## Detail
Some more context on why f933799c54b09210cb267963ff1dc431c7a9385a allows
for both flags to be passed now: I realized as part of
https://github.com/model-checking/kani/pull/3984 how when we call `cargo
rustc` for a `cargo kani` invocation, we don't pass `--reachability` to
dependencies to avoid running harnesses in them. The problem is that we
can't do the same for our cargo command to build the standard library,
since that uses `cargo build`, which does not have the same ability to
pass flags only to the final compiler invocation and not the
dependencies. So we end up passing `--reachability=AllFns` to the
dependencies of the standard library as well and generate automatic
harnesses for them. If we can pass both filter flags, we can run
commands like `kani autoharness --std --include-pattern core
--exclude-pattern miniz_oxide`, which will include functions from the
`core` crate while excluding functions in the `miniz_oxide` that just
happen to have the word "core" in them.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.