CI Formatting Fixes (#4028)
The majority of the changes in this PR were made automatically by
running the command specified in
5314a5d772cf4e8f152334dcbaa081336b631e02, so the actual diff to review
is much smaller than it initially appears.

1. If `--check` was provided to`kani-fmt.sh`, it would run `cargo fmt
--check` but not `rustfmt --check`. This meant that our regression would
catch formatting errors from `cargo fmt` but not `rustfmt` (it would
just format the problematic files, which isn't useful in CI--we want it
to fail instead). So add the `--check` flag to the `rustfmt` calls as
well.
2. Run clippy with `kani_sysroot` feature enabled as well to catch more
formatting errors.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.