Support `trait_upcasting` (#4001)
This adds support for the recently stabilized `trait_upcasting` feature.
Currently, when we cast between non-matching dyn traits, codegen
incorrectly reuses the old vtable, leading to confusing issues (#3998).

This reproduces the upcasting logic from upstream rustc to load the vptr
when performing the coercion.

Resolves #358
Resolves #3998

This was tested using the motivating example from the linked issue. I
will add further tests to make sure this feature is robust.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: clubby789 <jamie@osec.io>