Analyze unsafe code reachability (#4037)
Continuation of #3546.

From @celinval in #3546:

> Add call graph analysis to scanner in order to find the distance
between functions in a crate and unsafe functions.
> 
> For that, we build the crate call graph and collect the unsafe
functions. After that, do reverse BFS traversal from the unsafe
functions and store the distance to other functions. The result is
stored in a new csv file.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Celina G. Val <celinval@amazon.com>