Contracts/Stubs for multiple inherent impls: fix checking the generic args path for equality (#4051)
#3829 fixed an issue where we couldn't verify contracts or stubs for
types with multiple inherent impls. However, the logic to check for path
equality incorrectly assumed that there would only be one element in the
path before the generic argument, when in fact the path can be
arbitrarily long if there are modules involved. Change the logic to just
look at the last two elements, since we can expect those will be the
generic argument and the method name.

Before this PR, the new test would fail with this error:

```
error: Failed to resolve checking function NegativeNumber::<i32>::unchecked_mul because the generic arguments ::<i32> are invalid. The available implementations are: 
       num::negative::NegativeNumber::<i32>::unchecked_mul
       num::negative::NegativeNumber::<i16>::unchecked_mul
  --> /Users/cmzech/kani/tests/kani/FunctionContracts/multiple_inherent_impls.rs:57:5
   |
57 |     #[kani::proof_for_contract(NegativeNumber::<i32>::unchecked_mul)]
   |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   |
   = note: this error originates in the attribute macro `kani::proof_for_contract` (in Nightly builds, run with -Z macro-backtrace for more info)
```

because we'd try to compare
`::negative::NegativeNumber::<i32>::unchecked_mul` to
`::<i32>::unchecked_mul` and find they weren't equal.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.