Remove bool typedef (#4058)
This is a proposed alternative fix to #4052 where the new GCC version
(v15) switched from C17 to C23 by default and complains about errors in
https://github.com/model-checking/kani/blob/main/library/kani/kani_lib.c.

The first proposed fix in #4055 pinned the version to C17. This PR
instead removes the typedef that causes an error with C23.

Resolves #4052 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.