Fix the bug: Loop contracts are not composable with function contracts  (#3979)
This PR fixes the bug of loop contracts are not composable with function
contracts.
Previously, using loop-contract inside a function with contract may
result in unwinding the loop instead of contracting it.

Resolves #3910

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Carolyn Zech <cmzech@amazon.com>